% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uvpd.R
\name{.computeMatch}
\alias{.computeMatch}
\title{Title}
\usage{
.computeMatch(x, eps.mZ = 0.05, eps.rt = 0.5, treeDepth = 1,
  scanTypeFilter = "", ...)
}
\arguments{
\item{x}{}

\item{eps.mZ}{}

\item{eps.rt}{}

\item{treeDepth}{}

\item{...}{}
}
\value{

}
\description{
Title
}
