% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uvpd.R
\name{matchFragment}
\alias{matchFragment}
\title{Compute match between in-silico fragments and MS2 scans}
\usage{
matchFragment(MS2Scans, fragments, plot = FALSE, errorCutOff = 0.001,
  FUN = sum, ...)
}
\arguments{
\item{MS2Scans}{list of MS2 scans.}

\item{fragments}{\code{data.frame} of in-silico fragments.}

\item{plot}{\code{TRUE} or \code{FALSE}}

\item{errorCutOff}{mZ error cut-off.}

\item{FUN}{function used by aggregate.}

\item{...}{}
}
\value{
a aggregate data frame
}
\description{
Compute match between in-silico fragments and MS2 scans
}
