% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uvpd.R
\name{analyze}
\alias{analyze}
\alias{summary.uvpd}
\title{Perfom analyses run}
\usage{
analyze(rawfile, fragments, mZoffset = 1.007, errorCutOff = 1e-04)
}
\arguments{
\item{rawfile}{filepath to a Thermo Fisher rawfile}

\item{...}{passed to \code{uvpd:::.assignMatchedFragmentIons}, e.g. \code{FUN=sum}.}
}
\value{
returns a \code{data.frame} object
}
\description{
Perfom analyses run
}
\details{
INPUT:
\itemize{
\item in-silico fragment ion spectra of a given set of SMILES 
\item ThermoFisher raw file
}

STEP1: 
 \itemize{
 \item dermines APEX(MS1) of each molecule mass return rt for each given mass; 
}

STEP2: 
\itemize{
\item extract MS2 at a given rt and precursor window
}

STEP3: 
\itemize{
\item compute match between in-silico and MS2 fragment specs 
}

STEP4: 
\itemize{
\item aggregate (default \code{FUN=sum}) matches using
\code{intensity ~ mZ + type + SMILES + formula}.
where mZ are the matched fragment ion mass, type is fragment type,e.g., MH1P.
}
OUTPUT: 
\itemize{
\item a \code{data.frame} object having the column names
\code{c('mZ', 'type', 'SMILES', 'formula', 'intensity', 'rawfile',
  'scanTypeFilter', 'SMILES0', 'formula0', 'mass', 'n'}.
}
}
\examples{
library(uvpd)
load(file.path(system.file(package = 'uvpd'), "/extdata/fragments.RData"))

rawfiles <- scan(file.path(system.file(package = 'uvpd'),
  "/extdata/rawfiles.txt"), what = character())
  
rawfiles <- rawfiles[grepl("Castell|KWR|DB", rawfiles)]

\dontrun{
S1 <- analyze(rawfiles[20], fragments.treeDepth1)

S2 <- lapply(rawfiles[c(15:20)], analyze, fragments=fragments.treeDepth1)
}

\dontrun{
rawfile15 <- file.path(Sys.getenv('HOME'), "Downloads/CastellonStds_pos_HCD_20_35_60.raw")
rawfile20 <- file.path(Sys.getenv('HOME'), "Downloads/CastellonStds_pos_UVPD_50_300.raw")

S15 <- analyze(rawfile15, fragments = fragments.treeDepth1)
S20 <- analyze(rawfile20, fragments = fragments.treeDepth1)

S<-rbind(S15, S20)
}
}
\author{
Christian Panse <cp@fgcz.ethz.ch>, 2019
}
