% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uvpd.R
\name{analyze}
\alias{analyze}
\alias{summary.uvpd}
\title{Perfom analyses run}
\usage{
analyze(rawfile, fragments, ...)
}
\arguments{
\item{rawfile}{filepath to a Thermo Fisher rawfile}

\item{...}{passed to the other functions}
}
\value{
returns a \code{data.frame} object
}
\description{
Perfom analyses run
}
\details{
Input:
\itemize{
\item in-silicom fragment ion spectra of a given set of SMILES 
\item ThermoFisher raw file
}

STEP1: dermines APEX of each molecule mass

STEP2: compute match beween in-silico and MS2 specs

STEP3: aggregate matches

OUTPUT: a \code{data.frame} object having the column names
\code{c('mZ', 'type', 'SMILES', 'formula', 'intensity', 'rawfile', 'scanTypeFilter', 'SMILE0', 'mass', 'n'}.
}
\examples{
library(uvpd)
load(file.path(system.file(package = 'uvpd'), "/extdata/fragments.RData"))
load("~/__checkouts/R/uvpd/inst/extdata/uvpd-Meuse.RData") 
rawfiles <- scan(file.path(system.file(package = 'uvpd'),
  "/extdata/rawfiles.txt"), what = character())
  rawfiles <- rawfiles[grepl("Castell|KWR|DB", rawfiles)]

\dontrun{
S1 <- analyze(rawfiles[20], fragments.treeDepth1)

S2 <- mclapply(rawfiles[c(20,22,23()], analyze, fragments=fragments.treeDepth1)
}
}
\author{
Christian Panse <cp@fgcz.ethz.ch>
}
